Revert "Creates ops expressions"
This reverts commit cc6c9ea691351b10950be574845118d875dbe4ba.
